correct type hint for n_det_threshold

+++ ztfparsnip/create.py
@@ -305,7 +305,7 @@ class CreateLightcurves(object):
         sig_noise_cut: bool = True,
         delta_z: float = 0.1,
         SN_threshold: float = 5.0,
-        n_det_threshold: float = 5.0,
+        n_det_threshold: int = 5,
         subsampling_rate: float = 1.0,
         jd_scatter_sigma: float = 0.0,
         n: int | None = None,
